Fix inverter, rewriter, arith, optimizer, PB, GC, and datatype bugs
- expr_inverter: store m_is_var in parent class during set_is_var override
- horn_subsume_model_converter: translate m_delay_head/body in translate()
- arith_rewriter: return BR_FAILED for 0^(negative) instead of wrong 0
- opt_context: translate assumptions and values during context translation
- sat_clause: cap 8-bit m_inact_rounds at 255 to prevent wrap-around
- arith_solver: fix lpvar/theory_var type confusion for m_one_var/m_zero_var
- pb_pb: add Z3_fallthrough annotation for intentional case fallthrough
- smt_context: use uint64_t to prevent overflow in GC activity gradient
- theory_datatype: add null guards for dynamic_cast and get_enode results

+++ src/smt/smt_context.cpp
@@ -2828,7 +2828,7 @@ namespace smt {
                 // A clause is deleted if it has low activity and the number of unknowns is greater than a threshold.
                 // The activity threshold depends on how old the clause is.
                 unsigned act_threshold = m_fparams.m_old_clause_activity -
-                    (m_fparams.m_old_clause_activity - m_fparams.m_new_clause_activity) * ((i - start_at) / real_sz);
+                    static_cast<unsigned>(static_cast<uint64_t>(m_fparams.m_old_clause_activity - m_fparams.m_new_clause_activity) * (i - start_at) / real_sz);
                 if (cls->get_activity() < act_threshold) {
                     unsigned rel_threshold = (i >= new_first_idx ? m_fparams.m_new_clause_relevancy : m_fparams.m_old_clause_relevancy);
                     if (more_than_k_unassigned_literals(cls, rel_threshold)) {
